+++ packages/general/src/Buttons/TextButton.tsx
@@ -5,6 +5,7 @@ import {
 } from '@terraprisma/utils';
 
 import Button from './Button';
+import { createEffect } from 'solid-js';
 
 export interface TextButtonProps {
   /**