+++ packages/layout/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @terraprisma/layout
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @terraprisma/utils@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes