+++ packages/transitions/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @terraprisma/transitions
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @terraprisma/utils@0.1.2
+
 ## 0.0.2
 
 ### Patch Changes